requirements.txt
allure-pytest==2.13.5
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
iniconfig==1.1.1
pluggy==0.13.1
pytest==6.2.4
requests==2.26.0
selenium==3.141.0
urllib3==1.26.6
webdriver-manager==3.4.2


